Estruturando notificador de e-mail no modulo documental.
