+++ siga-ex/src/main/java/br/gov/jfrj/siga/ex/bl/ExBL.java
@@ -218,10 +218,12 @@ import br.gov.jfrj.siga.ex.model.enm.ExTipoDeVinculo;
 import br.gov.jfrj.siga.ex.service.ExService;
 import br.gov.jfrj.siga.ex.util.DatasPublicacaoDJE;
 import br.gov.jfrj.siga.ex.util.ExMovimentacaoRecebimentoComparator;
+import br.gov.jfrj.siga.ex.util.Email;
 import br.gov.jfrj.siga.ex.util.FuncoesEL;
 import br.gov.jfrj.siga.ex.util.GeradorRTF;
 import br.gov.jfrj.siga.ex.util.MascaraUtil;
 import br.gov.jfrj.siga.ex.util.Notificador;
+import br.gov.jfrj.siga.ex.util.Notificar;
 import br.gov.jfrj.siga.ex.util.ProcessadorHtml;
 import br.gov.jfrj.siga.ex.util.ProcessadorModelo;
 import br.gov.jfrj.siga.ex.util.ProcessadorModeloFreemarker;
@@ -1695,13 +1697,8 @@ public class ExBL extends CpBL {
 					Ex.getInstance().getBL().gravar(cadastrante, titular, mov.getLotaTitular(), doc);
 				}
 				
-				List<DpPessoa> cossignatarios = doc.getCosignatarios();  
-				for (DpPessoa cossignatario : cossignatarios) {
-					if (Cp.getInstance().getConf().podeUtilizarServicoPorConfiguracao(cossignatario, cossignatario.getLotacao(), 
-							CpServicosNotificacaoPorEmail.COSSIG.getChave()) && !doc.isAssinadoPeloSubscritorComTokenOuSenha()) {
-						enviarEmailAoCossignatario(cossignatario, cadastrante, doc.getSigla());
-					}
-				}
+//				notificar = new Notificar();
+//				notificar.cossignatario(doc, cadastrante);
 				
 				gravarMovimentacao(mov);
 	
@@ -1964,13 +1961,8 @@ public class ExBL extends CpBL {
 					Ex.getInstance().getBL().gravar(cadastrante, titular, mov.getLotaTitular(), doc);
 				}
 				
-				List<DpPessoa> cossignatarios = doc.getCosignatarios();  
-				for (DpPessoa cossignatario : cossignatarios) {
-					if (Cp.getInstance().getConf().podeUtilizarServicoPorConfiguracao(cossignatario, cossignatario.getLotacao(), 
-							CpServicosNotificacaoPorEmail.COSSIG.getChave()) && !doc.isAssinadoPeloSubscritorComTokenOuSenha()) {
-						enviarEmailAoCossignatario(cossignatario, cadastrante, doc.getSigla());
-					}
-				} 
+//				notificar = new Notificar();
+//				notificar.cossignatario(doc, cadastrante);
 				
 				gravarMovimentacao(mov);
 	
@@ -3169,7 +3161,7 @@ public class ExBL extends CpBL {
 			doc.setNumPaginas(doc.getContarNumeroDePaginas());
 			
 			dao().gravar(doc);
-
+			
 			if (doc.getExFormaDocumento().getExTipoFormaDoc().isExpediente()) {
 				for (final ExVia via : setVias) {
 					Integer numVia = null;
@@ -3195,10 +3187,8 @@ public class ExBL extends CpBL {
 			if (doc.getExMobilAutuado() != null)
 				juntarAoDocumentoAutuado(cadastrante, lotaCadastrante, doc, null, cadastrante);
 			
-			if (Cp.getInstance().getConf().podeUtilizarServicoPorConfiguracao(doc.getSubscritor(), 
-					doc.getSubscritor().getLotacao(), CpServicosNotificacaoPorEmail.RESPASS.getChave())) { 
-				enviarEmailResponsavelPelaAssinatura(cadastrante, doc.getSubscritor(), doc.getSigla());
-			}
+//			notificar = new Notificar();
+//			notificar.responsavelPelaAssinatura(doc, cadastrante);
 			
 			return s;
 		} catch (final Exception e) {
@@ -4065,10 +4055,9 @@ public class ExBL extends CpBL {
 			mov.getExMobil().setDnmDataUltimaMovimentacaoNaoCancelada(movUlt.getDtIniMov());
 			dao().gravar(mov.getExMobil());
 		}
-
-		if (mov.getExTipoMovimentacao() != ExTipoDeMovimentacao.CANCELAMENTO_DE_MOVIMENTACAO) {
-			Notificador.notificarDestinariosEmail(mov, Notificador.TIPO_NOTIFICACAO_GRAVACAO);
-		}
+		
+		Notificar notificar = new Notificar(mov.getCadastrante());
+		notificar.enviar(mov);
 		
 		if (SigaMessages.isSigaSP()) {
 			if (mov.getExTipoMovimentacao() == ExTipoDeMovimentacao.SOLICITACAO_DE_ASSINATURA &&
@@ -5348,58 +5337,8 @@ public class ExBL extends CpBL {
 						listaMovimentacao.addAll(m.doc().getMobilGeral()
 								.getMovsNaoCanceladas(ExTipoDeMovimentacao.RESTRINGIR_ACESSO));
 						
-						Set<ExMobil> exMobils = mov.getExDocumento().getExMobilSet();
-						Set<DpPessoa> pessoasLota = mov.getLotaResp().getDpPessoaLotadosSet();
-						Set<CpMarcador> marcas = new HashSet<>();
-						StringJoiner marcasDoDoc = new StringJoiner(", ");  
-
-						if (mov.getResp() != null) { 
-							if (Cp.getInstance().getConf().podeUtilizarServicoPorConfiguracao(mov.getResp(), 
-									mov.getResp().getLotacao(), CpServicosNotificacaoPorEmail.DOCMARC.getChave())) {
-								for (ExMobil exMobil: exMobils) {
-									for (ExMarca exMarca: exMobil.getExMarcaSet()) {
-										if (exMarca.getCpMarcador().getIdFinalidade().getIdTpMarcador() == CpTipoMarcadorEnum.TIPO_MARCADOR_LOTACAO 
-												|| exMarca.getCpMarcador().getIdFinalidade().getIdTpMarcador() == CpTipoMarcadorEnum.TIPO_MARCADOR_GERAL) {
-											marcas.add(exMarca.getCpMarcador());
-										}
-									}
-								}
-								if (marcasDoDoc.length() == 0) {
-									marcas.forEach(marc -> {
-										marcasDoDoc.add(marc.getDescrMarcador());
-									});
-								}
-								enviarEmailAoTramitarDocMarcado(mov.getResp(), mov.getTitular(), mov.getExDocumento().getSigla(), marcasDoDoc + "");	
-							}
-							if (Cp.getInstance().getConf().podeUtilizarServicoPorConfiguracao(mov.getResp(), 
-										mov.getResp().getLotacao(), CpServicosNotificacaoPorEmail.DOCTUSU.getChave()))  
-								enviarEmailAoTramitarDocParaUsuario(mov.getResp(), mov.getTitular(), mov.getExDocumento().getSigla());
-						}
-
-						if (mov.getResp() == null) {
-							for (DpPessoa pessoa: pessoasLota) {
-								if (Cp.getInstance().getConf().podeUtilizarServicoPorConfiguracao(pessoa, 
-										pessoa.getLotacao(), CpServicosNotificacaoPorEmail.DOCMARC.getChave())) { 
-									for (ExMobil exMobil: exMobils) {
-										for (ExMarca exMarca: exMobil.getExMarcaSet()) {
-											if (exMarca.getCpMarcador().getIdFinalidade().getIdTpMarcador() == CpTipoMarcadorEnum.TIPO_MARCADOR_LOTACAO 
-													|| exMarca.getCpMarcador().getIdFinalidade().getIdTpMarcador() == CpTipoMarcadorEnum.TIPO_MARCADOR_GERAL) {
-												marcas.add(exMarca.getCpMarcador());
-											}
-										}
-									}
-									if (marcasDoDoc.length() == 0) {
-										marcas.forEach(marc -> {
-											marcasDoDoc.add(marc.getDescrMarcador());
-										}); 
-									} 
-									enviarEmailAoTramitarDocMarcado(pessoa, mov.getTitular(), mov.getExDocumento().getSigla(), marcasDoDoc + "");		
-								} 
-								if (Cp.getInstance().getConf().podeUtilizarServicoPorConfiguracao(pessoa, 
-										pessoa.getLotacao(), CpServicosNotificacaoPorEmail.DOCTUN.getChave())) 
-										enviarEmailAoTramitarDocParaUsuariosDaUnidade(mov.getLotaResp(), pessoa, mov.getExDocumento().getSigla());
-							}
-						}
+//						notificar = new Notificar();
+//						notificar.usuarioDiretamenteOuPelaUnidade(mov);
 						
 						if (!listaMovimentacao.isEmpty()) {
 							List<ExDocumento> listaDocumentos = new ArrayList<ExDocumento>();
@@ -8492,181 +8431,6 @@ public class ExBL extends CpBL {
 			throw new AplicacaoException("Erro ao gravar link público do documento", ExTipoDeMovimentacao.GERAR_LINK_PUBLICO_PROCESSO.getId(), e);
 		}
 
-	}
-	
-	private String docTramitadoParaUsuario(DpPessoa destinatario, DpPessoa cadastrante, String siglaDoc) {		 
-		String conteudo = "";
-		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/doc-tramitado-para-usuario.html"),StandardCharsets.UTF_8))) {			
-			String str;
-			while((str = bfr.readLine()) != null) {
-				conteudo += str;
-			}
-			conteudo = conteudo
-					.replace("${url}", Prop.get("/siga.base.url"))
-					.replace("${logo}", Prop.get("/siga.email.logo"))
-					.replace("${titulo}", Prop.get("/siga.email.titulo"))
-					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
-					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
-					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
-					.replace("${siglaCadastrante}", cadastrante.getSigla())
-					.replace("${siglaDoc}", siglaDoc);
-
-			return conteudo;
-
-		} catch (IOException e) {
-			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
-		}	
-	}
-
-	private String docMarcadoTramitadoParaUsuario(DpPessoa destinatario, DpPessoa cadastrante, String docSigla, String marcador) {		 
-		String conteudo = "";
-		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/doc-marcado-tramitado-para-unidade.html"),StandardCharsets.UTF_8))) {			
-			String str;
-			while((str = bfr.readLine()) != null) {
-				conteudo += str;
-			}
-			conteudo = conteudo
-					.replace("${url}", Prop.get("/siga.base.url"))
-					.replace("${logo}", Prop.get("/siga.email.logo"))
-					.replace("${titulo}", Prop.get("/siga.email.titulo"))
-					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
-					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
-					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
-					.replace("${siglaCadastrante}", cadastrante.getSigla())
-					.replace("${docSigla}", docSigla)
-					.replace("${marcador}", marcador); 
-
-			return conteudo;
-
-		} catch (IOException e) {
-			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
-		}	
-	}
-
-	private String responsavelPelaAssinatura(DpPessoa destinatario, DpPessoa cadastrante, String siglaDoc) {		 
-		String conteudo = "";
-		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/usuario-responsavel-pela-assinatura.html"),StandardCharsets.UTF_8))) {			
-			String str;
-			while((str = bfr.readLine()) != null) {
-				conteudo += str;
-			}
-			conteudo = conteudo
-					.replace("${url}", Prop.get("/siga.base.url"))
-					.replace("${logo}", Prop.get("/siga.email.logo"))
-					.replace("${titulo}", Prop.get("/siga.email.titulo"))
-					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
-					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
-					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
-					.replace("${siglaCadastrante}", cadastrante.getSigla())
-					.replace("${siglaDoc}", siglaDoc);
-
-			return conteudo;
-
-		} catch (IOException e) {
-			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
-		}	
-	}
-
-	private String incluirCossignatario(DpPessoa destinatario, DpPessoa cadastrante, String siglaDoc) {		 
-		String conteudo = "";
-		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/incluido-como-cossignatario.html"),StandardCharsets.UTF_8))) {			
-			String str;
-			while((str = bfr.readLine()) != null) {
-				conteudo += str;
-			}
-			conteudo = conteudo
-					.replace("${url}", Prop.get("/siga.base.url"))
-					.replace("${logo}", Prop.get("/siga.email.logo"))
-					.replace("${titulo}", Prop.get("/siga.email.titulo"))
-					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
-					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
-					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
-					.replace("${siglaCadastrante}", cadastrante.getSigla()) 
-					.replace("${siglaDoc}", siglaDoc);
-
-			return conteudo;
-
-		} catch (IOException e) {
-			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
-		}	
-	}
-
-	private String docTramitadoParaUnidade(DpPessoa destinatario, DpLotacao lotacao, String docSigla) {		 
-		String conteudo = "";
-		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/doc-tramitado-para-unidade.html"),StandardCharsets.UTF_8))) {			
-			String str;
-			while((str = bfr.readLine()) != null) {
-				conteudo += str;
-			}
-			conteudo = conteudo
-					.replace("${url}", Prop.get("/siga.base.url"))
-					.replace("${logo}", Prop.get("/siga.email.logo"))
-					.replace("${titulo}", Prop.get("/siga.email.titulo"))
-					.replace("${docSigla}", docSigla)
-					.replace("${lotaDesc}", lotacao.getDescricao())
-					.replace("${lotaSigla}", lotacao.getSigla());
-
-			return conteudo;
-
-		} catch (IOException e) {
-			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
-		}	
-	}
-	
-	public void enviarEmailAoTramitarDocParaUsuario(DpPessoa pessoaDest, DpPessoa titular, String sigla) {
-		String assunto = "Documento tramitado para " + pessoaDest.getDescricao();
-		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
-		String conteudoHTML = docTramitadoParaUsuario(pessoaDest, titular, sigla);
-		try {
-			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
-		} catch (Exception e) {
-			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
-		}
-	} 
-
-	public void enviarEmailResponsavelPelaAssinatura(DpPessoa pessoaDest, DpPessoa titular, String sigla) {
-		String assunto = "Responsável pela assinatura: " + pessoaDest.getDescricao();
-		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
-		String conteudoHTML = responsavelPelaAssinatura(pessoaDest, titular, sigla);
-		try {
-			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
-		} catch (Exception e) {
-			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
-		}
-	}
-
-	public void enviarEmailAoTramitarDocMarcado(DpPessoa pessoaDest, DpPessoa titular, String sigla, String marcador) {
-		String assunto = "Documento tramitado para " + pessoaDest.getDescricao();
-		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
-		String conteudoHTML = docMarcadoTramitadoParaUsuario(pessoaDest, titular, sigla, marcador);
-		try {
-			if (!marcador.equals(""))
-			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
-		} catch (Exception e) {
-			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
-		}
-	}
-
-	public void enviarEmailAoTramitarDocParaUsuariosDaUnidade(DpLotacao lotaDest, DpPessoa pessoa, String sigla) {
-		String assunto = "Documento tramitado para unidade " + lotaDest.getDescricao();
-		String[] destinanarios = { pessoa.getEmailPessoaAtual() };
-		String conteudoHTML = docTramitadoParaUnidade(pessoa, lotaDest, sigla);
-		try {
-			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
-		} catch (Exception e) {
-			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
-		}
-	}
-
-	public void enviarEmailAoCossignatario(DpPessoa pessoaDest, DpPessoa titular, String sigla) {
-		String assunto = "Documento tramitado para " + pessoaDest.getDescricao();
-		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
-		String conteudoHTML = incluirCossignatario(pessoaDest, titular, sigla);
-		try {
-			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
-		} catch (Exception e) {
-			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
-		}
 	}
 
 }
