+++ siga-ex/src/main/java/br/gov/jfrj/siga/ex/util/Email.java
@@ -1,0 +1,196 @@
+package br.gov.jfrj.siga.ex.util;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+
+import br.gov.jfrj.siga.base.AplicacaoException;
+import br.gov.jfrj.siga.base.Correio;
+import br.gov.jfrj.siga.base.Prop;
+import br.gov.jfrj.siga.dp.DpLotacao;
+import br.gov.jfrj.siga.dp.DpPessoa;
+
+/**
+ * Essa classe tem a responsabilidade de montar e enviar o e-mail.
+ * 
+ */
+
+public class Email implements EnviarEmail {
+
+	public void enviarAoTramitarDocParaUsuario(DpPessoa pessoaDest, DpPessoa titular, String sigla) {
+		String assunto = "Documento tramitado para " + pessoaDest.getDescricao();
+		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
+		String conteudoHTML = docTramitadoParaUsuario(pessoaDest, titular, sigla);
+		try {
+			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
+		} catch (Exception e) {
+			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
+		}
+	} 
+
+	public void enviarAoResponsavelPelaAssinatura(DpPessoa pessoaDest, DpPessoa titular, String sigla) {
+		String assunto = "Responsável pela assinatura: " + pessoaDest.getDescricao();
+		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
+		String conteudoHTML = responsavelPelaAssinatura(pessoaDest, titular, sigla);
+		try {
+			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
+		} catch (Exception e) {
+			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
+		}
+	}
+
+	public void enviarAoTramitarDocMarcado(DpPessoa pessoaDest, DpPessoa titular, String sigla, String marcador) {
+		String assunto = "Documento tramitado para " + pessoaDest.getDescricao();
+		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
+		String conteudoHTML = docMarcadoTramitadoParaUsuario(pessoaDest, titular, sigla, marcador);
+		try {
+			if (!marcador.equals(""))
+			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
+		} catch (Exception e) {
+			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
+		}
+	}
+
+	public void enviarAoTramitarDocParaUsuariosDaUnidade(DpLotacao lotaDest, DpPessoa pessoa, String sigla) {
+		String assunto = "Documento tramitado para unidade " + lotaDest.getDescricao();
+		String[] destinanarios = { pessoa.getEmailPessoaAtual() };
+		String conteudoHTML = docTramitadoParaUnidade(pessoa, lotaDest, sigla);
+		try {
+			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
+		} catch (Exception e) {
+			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
+		}
+	}
+
+	public void enviarAoCossignatario(DpPessoa pessoaDest, DpPessoa titular, String sigla) {
+		String assunto = "Documento tramitado para " + pessoaDest.getDescricao();
+		String[] destinanarios = { pessoaDest.getEmailPessoaAtual() };
+		String conteudoHTML = incluirCossignatario(pessoaDest, titular, sigla);
+		try {
+			Correio.enviar(null,destinanarios, assunto, "", conteudoHTML);
+		} catch (Exception e) {
+			throw new AplicacaoException("Ocorreu um erro durante o envio do email", 0, e);
+		}
+	}
+	
+	private String docTramitadoParaUsuario(DpPessoa destinatario, DpPessoa cadastrante, String siglaDoc) {		 
+		String conteudo = "";
+		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/doc-tramitado-para-usuario.html"),StandardCharsets.UTF_8))) {			
+			String str;
+			while((str = bfr.readLine()) != null) {
+				conteudo += str;
+			}
+			conteudo = conteudo
+					.replace("${url}", Prop.get("/siga.base.url"))
+					.replace("${logo}", Prop.get("/siga.email.logo"))
+					.replace("${titulo}", Prop.get("/siga.email.titulo"))
+					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
+					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
+					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
+					.replace("${siglaCadastrante}", cadastrante.getSigla())
+					.replace("${siglaDoc}", siglaDoc);
+
+			return conteudo;
+
+		} catch (IOException e) {
+			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
+		}	
+	}
+
+	private String docMarcadoTramitadoParaUsuario(DpPessoa destinatario, DpPessoa cadastrante, String docSigla, String marcador) {		 
+		String conteudo = "";
+		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/doc-marcado-tramitado-para-unidade.html"),StandardCharsets.UTF_8))) {			
+			String str;
+			while((str = bfr.readLine()) != null) {
+				conteudo += str;
+			}
+			conteudo = conteudo
+					.replace("${url}", Prop.get("/siga.base.url"))
+					.replace("${logo}", Prop.get("/siga.email.logo"))
+					.replace("${titulo}", Prop.get("/siga.email.titulo"))
+					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
+					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
+					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
+					.replace("${siglaCadastrante}", cadastrante.getSigla())
+					.replace("${docSigla}", docSigla)
+					.replace("${marcador}", marcador); 
+
+			return conteudo;
+
+		} catch (IOException e) {
+			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
+		}	
+	}
+
+	private String responsavelPelaAssinatura(DpPessoa destinatario, DpPessoa cadastrante, String siglaDoc) {		 
+		String conteudo = "";
+		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/usuario-responsavel-pela-assinatura.html"),StandardCharsets.UTF_8))) {			
+			String str;
+			while((str = bfr.readLine()) != null) {
+				conteudo += str;
+			}
+			conteudo = conteudo
+					.replace("${url}", Prop.get("/siga.base.url"))
+					.replace("${logo}", Prop.get("/siga.email.logo"))
+					.replace("${titulo}", Prop.get("/siga.email.titulo"))
+					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
+					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
+					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
+					.replace("${siglaCadastrante}", cadastrante.getSigla())
+					.replace("${siglaDoc}", siglaDoc);
+
+			return conteudo;
+
+		} catch (IOException e) {
+			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
+		}	
+	}
+
+	private String incluirCossignatario(DpPessoa destinatario, DpPessoa cadastrante, String siglaDoc) {		 
+		String conteudo = "";
+		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/incluido-como-cossignatario.html"),StandardCharsets.UTF_8))) {			
+			String str;
+			while((str = bfr.readLine()) != null) {
+				conteudo += str;
+			}
+			conteudo = conteudo
+					.replace("${url}", Prop.get("/siga.base.url"))
+					.replace("${logo}", Prop.get("/siga.email.logo"))
+					.replace("${titulo}", Prop.get("/siga.email.titulo"))
+					.replace("${nomeDestinatario}", destinatario.getNomePessoa())  
+					.replace("${siglaDestinatario}", destinatario.getSiglaCompleta())
+					.replace("${nomeCadastrante}", cadastrante.getNomePessoa())
+					.replace("${siglaCadastrante}", cadastrante.getSigla()) 
+					.replace("${siglaDoc}", siglaDoc);
+
+			return conteudo;
+
+		} catch (IOException e) {
+			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
+		}	
+	}
+
+	private String docTramitadoParaUnidade(DpPessoa destinatario, DpLotacao lotacao, String docSigla) {		 
+		String conteudo = "";
+		try (BufferedReader bfr = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("/templates/email/doc-tramitado-para-unidade.html"),StandardCharsets.UTF_8))) {			
+			String str;
+			while((str = bfr.readLine()) != null) {
+				conteudo += str;
+			}
+			conteudo = conteudo
+					.replace("${url}", Prop.get("/siga.base.url"))
+					.replace("${logo}", Prop.get("/siga.email.logo"))
+					.replace("${titulo}", Prop.get("/siga.email.titulo"))
+					.replace("${docSigla}", docSigla)
+					.replace("${lotaDesc}", lotacao.getDescricao())
+					.replace("${lotaSigla}", lotacao.getSigla());
+
+			return conteudo;
+
+		} catch (IOException e) {
+			throw new AplicacaoException("Erro ao montar e-mail para enviar ao usuário " + destinatario.getNomePessoa());
+		}	
+	}
+	
+}
