+++ siga-vraptor-module/src/main/resources/META-INF/tags/menuprincipal.tag
@@ -204,8 +204,10 @@
 					<li><a class="dropdown-item"
 						href="/siga/app/usuario/trocar_email">Trocar Email do Usuário</a></li>
 				</c:if>
-				<li><a class="dropdown-item"
-						href="/siga/app/notificarPorEmail/listar">Receber notificações por email</a></li>
+				<c:if test="${(f:podeUtilizarServicoPorConfiguracao(titular,lotaTitular,'SIGA;CEMAIL: Módulo de notificação por email'))}">
+					<li><a class="dropdown-item"
+							href="/siga/app/notificarPorEmail/listar">Receber notificações por email</a></li>
+				</c:if> 
 				<c:if test="${f:podeUtilizarSegundoFatorPin(cadastrante,cadastrante.lotacao) }">
 					<li class="dropdown-submenu"><a href="javascript:void(0);" class="dropdown-item dropdown-toggle">Gerenciar PIN</a>
 						<ul class="dropdown-menu">
